refactor(frontend): refactor AlterSchemaPrepForm ready for tenant mode in the future
