+++ frontend/src/components/AlterSchemaPrepForm/ProjectTenantView.vue
@@ -1,0 +1,11 @@
+<template>
+  <div>TODO: this will NOT be visible for now</div>
+</template>
+
+<script lang="ts">
+import { defineComponent } from "vue";
+
+export default defineComponent({
+  name: "ProjectTenantView",
+});
+</script>
