+++ frontend/src/components/AlterSchemaPrepForm/index.ts
@@ -1,0 +1,3 @@
+import AlterSchemaPrepForm from "./AlterSchemaPrepForm.vue";
+
+export default AlterSchemaPrepForm;
