+++ frontend/src/types/common.ts
@@ -197,6 +197,7 @@ export const unknown = (
     memberList: [],
     workflowType: "UI",
     visibility: "PUBLIC",
+    tenantMode: "DISABLED",
   };
 
   const UNKNOWN_PROJECT_HOOK: ProjectWebhook = {
@@ -548,6 +549,7 @@ export const empty = (
     memberList: [],
     workflowType: "UI",
     visibility: "PUBLIC",
+    tenantMode: "DISABLED",
   };
 
   const EMPTY_PROJECT_HOOK: ProjectWebhook = {
